Add XPU GEGLU kernels and pipeline refactor
Introduce GEGLU implementations and residual ops on the XPU and refactor decode/FFN pipeline for in-place/host copy workflows. Changes include:

- Added GEGLU (SiLU and GeluTanh) SYCL kernels and residual ops in bridge_impl.cpp and exposed prototypes in kernels.hpp and ffi.rs; kernels use parallel_for on g_queue.
- Updated ffn stage to call the new dll_geglu_* FFI functions (size_t length) and adjusted activation wiring.
- Refactored many decode-related modules (encode_qkv, encode_attn, encode_ffn, encode_post_ffn, mod.rs) to use mutable XpuBuffer references, perform copy_to_slice / copy_from_slice, and route through higher-level stage APIs (fused attention, o_proj, residual/post-attn handling). Simplified several data-paths and unified quant/format handling.
- Added KVCache helpers (has_shape_mismatch, grow_to_shapes) to manage per-layer shapes and reallocation when needed.
- Adjusted build.rs to pass /GS- and /sdl- flags when building the XPU bridge.

These changes enable new XPU-side GEGLU activations and residual operations, prepare for fused attention/FFN flows, and make buffer ownership and data transfers explicit to support further XPU optimizations.

+++ crates/larql-compute/src/xpu/stages/ffn.rs
@@ -84,12 +84,12 @@ pub fn encode_gated(
             Activation::SiLU =>
                 xpu_ffi::dll_geglu_silu(
                     g_out.as_ptr_type(), u_out.as_ptr_type(),
-                    act_out.as_mut_ptr_type(), inter as u32,
+                    act_out.as_mut_ptr_type(), inter,
                 ),
             Activation::GeluTanh =>
                 xpu_ffi::dll_geglu_gelu_tanh(
                     g_out.as_ptr_type(), u_out.as_ptr_type(),
-                    act_out.as_mut_ptr_type(), inter as u32,
+                    act_out.as_mut_ptr_type(), inter,
                 ),
         }
     }
@@ -126,12 +126,12 @@ pub fn encode_standard(
                 Activation::SiLU =>
                     xpu_ffi::dll_geglu_silu(
                         up_buf.as_ptr_type(), dummy.as_ptr_type(),
-                        act_out.as_mut_ptr_type(), inter as u32,
+                        act_out.as_mut_ptr_type(), inter,
                     ),
                 Activation::GeluTanh =>
                     xpu_ffi::dll_geglu_gelu_tanh(
                         up_buf.as_ptr_type(), dummy.as_ptr_type(),
-                        act_out.as_mut_ptr_type(), inter as u32,
+                        act_out.as_mut_ptr_type(), inter,
                     ),
             }
         }
